Add basic client
Add a basic client... however the plex API along with however it's
internally tracking "servers" attached to them is pretty messy. You
also need to connect to plex.tv and it grabs multiple IPs that it
thinks your account lives at... some of which are gateway ips...
all in all plex API is a huge piece of shit and it's likely not
worth trying to nicely configure a plex server.